% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{pkg_fn_summary}
\alias{pkg_fn_summary}
\title{Return a table with package functions call summary}
\usage{
pkg_fn_summary(path = ".", igraph_obj = NULL)
}
\arguments{
\item{path}{package path}

\item{igraph_obj}{igraph object for function calls dependencies returned by `create_package_igraph()`}
}
\value{
a table with package functions call summary
}
\description{
Return a table with package functions call summary
}
