% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview_init.R
\name{pkgreview_create}
\alias{pkgreview_create}
\alias{pkgreview_init}
\title{Create and initialise package review}
\usage{
pkgreview_create(pkg_repo, review_dir = ".")

pkgreview_init(pkg_repo, review_dir = ".")
}
\arguments{
\item{pkg_repo}{character string of the repo owner and name in the form of
`"owner/repo"`.}

\item{review_dir}{directory in which to setup review project and source package
source code.}
}
\value{
setup review project with templates
}
\description{
Create a review project using `pkgreview_create`. Wait till your workspace has
been switched to pkg review project dircetory and initialise review using
`pkgreview_init`
}
\examples{
\dontrun{
pkgreview_init(pkg_repo = "cboettig/rdflib")
}
}
