% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview.R
\name{pkgreview_getdata}
\alias{pkgreview_getdata}
\title{pkgreview_getdata}
\usage{
pkgreview_getdata(
  pkg_dir = NULL,
  pkg_repo,
  template = c("review", "editor"),
  issue_no = NULL
)
}
\arguments{
\item{pkg_dir}{path to package source directory, cloned from github. Defaults
to the package source code directory in the review parent.}

\item{pkg_repo}{character string of the repo owner and name in the form of
\code{"owner/repo"}.}

\item{template}{character string, one of \code{review} or \code{editor}.}

\item{issue_no}{integer. Issue number of the pkg review in the rOpenSci \href{https://github.com/ropensci/software-review/issues}{\code{software-review} repository}.
If \code{NULL} (default), the issue number is extracted from the rOpenSci \strong{Under Review} badge on the pkg repository README.
Supplying an integer to \code{issue_no} overrides this behaviour and can be useful if a badge has not been added to the README yet.}
}
\value{
a list of package metadata
}
\description{
get package metadata from package source code.
}
\examples{
\dontrun{
# run from within a pkgreviewr project with the package source code in a
sibling directory
pkgreview_getdata("../rdflib")
}
}
