% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview_init.R
\name{pkgreview_init}
\alias{pkgreview_init}
\title{Initialise package review}
\usage{
pkgreview_init(pkg_repo = "cboettig/rdflib", review_dir = ".",
  issue_url = "https://github.com/ropensci/onboarding/issues/169")
}
\arguments{
\item{pkg_repo}{character string of the repo owner and name in the form of
`"owner/repo"`.}

\item{review_dir}{directory in which to setup review project and source package
source code.}

\item{issue_url}{url of issue # in the rOpenSci onboarding
<https://github.com/ropensci/onboarding/issues>}
}
\value{
setup review project with templates
}
\description{
Initialise package review
}
\examples{
\dontrun{
pkgreview_init(pkg_repo = "cboettig/rdflib")
}
}
