% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-utils.R
\name{pkgreview_print_source}
\alias{pkgreview_print_source}
\title{Print review package function source code}
\usage{
pkgreview_print_source(pkgname)
}
\arguments{
\item{pkgname}{character string. review package name}
}
\value{
prints out function source code for all exported functions.
}
\description{
Print review package function source code
}
\examples{
\dontrun{
library("pkgreviewr")
pkgreview_print_source("pkgreviewr")
}
}
