% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview_init.R
\name{pkgreview_readme_rmd}
\alias{pkgreview_readme_rmd}
\alias{pkgreview_readme_md}
\title{Create README files.}
\usage{
pkgreview_readme_rmd(pkg_dir, issue_url, open = interactive())

pkgreview_readme_md(pkg_dir, issue_url, open = interactive())
}
\arguments{
\item{pkg_dir}{package directory}

\item{issue_url}{}

\item{open}{open on}
}
\description{
Creates skeleton README files with sections for
\itemize{
\item a high-level description of the package and its goals
\item R code to install from GitHub, if GitHub usage detected
\item a basic example
}
Use `Rmd` if you want a rich intermingling of code and data. Use
`md` for a basic README. `README.Rmd` will be automatically
added to `.Rbuildignore`. The resulting README is populated with default
YAML frontmatter and R fenced code blocks (`md`) or chunks (`Rmd`).
}
\examples{
\dontrun{
use_readme_rmd()
use_readme_md()
}
}
