% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreviewr-package.R
\docType{package}
\name{pkgreviewr-package}
\alias{pkgreviewr}
\alias{pkgreviewr-package}
\title{pkgreviewr: rOpenSci package review project template}
\description{
Creates all files necessary to complete the review, prepopulated
with review package specific metadata. Also clones the source code of the package
under review.
}
\keyword{internal}
