% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-templates.R
\name{render_request}
\alias{render_request}
\title{Render request email body}
\usage{
render_request()
}
\value{
renders \code{request.Rmd} using parameters provided.
}
\description{
Launches an interactive input browser tab to complete required parameters:
\itemize{
\item \code{reviewer_first_name}: reviewers first name
\item \code{banter}: character string of custom greeting message
\item \code{JOSS}: logical, whether review includes submission to JOSS
}
}
\examples{
\dontrun{
render_request()
}
}
