% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_args.R
\name{rev_args}
\alias{rev_args}
\title{Extracts the default arguments and checks for consistency}
\usage{
rev_args(path = ".", exported_only = FALSE)
}
\arguments{
\item{path}{Name of a source R package tar archive file,
or path to the folder of an R package.}

\item{exported_only}{Logical. Whether to check only the exported functions
or allt he functions detected by \link{rev_calls}.}
}
\value{
A two element list with \code{arg_df} and \code{arg_map}.
The \code{arg_df} is a data.frame with one element per argument that
has the name of the argument, the number of functions where it's used,
whether the default is consistent across all functions and percent
of consistency.
\code{arg_map} is a logical matrix with the functions in the rows and the
arguments in the columns. It specifies which functions use which arguments.
}
\description{
From a package, get all the default arguments for the functions of said
package. Then check how many times each argument is used and whether
the default value for teh argument is consistent across all functions
where it's used. This function also returns a matrix indicating each function
where the arguments are used.
}
\examples{
\dontrun{
## If your current working directory is the home of a package, use:
rev_args('.')
}

}
\author{
Leonardo Collado-Torres \url{https://github.com/lcolladotor}
}
