% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev_calls}
\alias{rev_calls}
\title{Create a dataframe of functions that a called and called by}
\usage{
rev_calls(path = ".", igraph_obj = NULL)
}
\arguments{
\item{path}{package path}

\item{igraph_obj}{igraph object for function calls dependencies returned by `create_package_igraph()`}
}
\value{
A dataframe of functions in and used by the package and the number of times they are called and called by
}
\description{
Create a dataframe of all functions in and used by the package
}
