% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_dependency_usage.R
\name{rev_dependency_usage}
\alias{rev_dependency_usage}
\title{Review dependency usage: Count used functions from external packages.}
\usage{
rev_dependency_usage(path = ".", include_base = FALSE)
}
\arguments{
\item{path}{Name of a source R package tar archive file,
or path to the folder of an R package.}

\item{include_base}{Whether to include functions from the
`base` package.}
}
\value{
A dataframe.
}
\description{
Review dependency usage: Count used functions from external packages.
}
\examples{
\dontrun{
rev_dependency_usage()
}
}
