% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-templates.R
\name{use_reviewtmpl}
\alias{use_reviewtmpl}
\title{Create review sumbission}
\usage{
use_reviewtmpl()
}
\value{
writes a `pkgreview.md` template submission file in the review project root.
}
\description{
Clone an up to date ropensci onboarding review template.
}
\examples{
\dontrun{
use_reviewtmpl()
}
}
