% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview_init.R
\name{pkgreview_create}
\alias{pkgreview_create}
\alias{pkgreview_init}
\title{Create and initialise package review}
\usage{
pkgreview_create(pkg_repo, review_parent = ".", open = TRUE)

pkgreview_init(pkg_repo, review_dir = here::here(), open = interactive())
}
\arguments{
\item{pkg_repo}{character string of the repo owner and name in the form of
`"owner/repo"`.}

\item{review_parent}{directory in which to setup review project and source package
source code.}

\item{open}{Open the newly created project/file for editing? Happens in RStudio, if
applicable, or via [utils::file.edit()] otherwise.}

\item{review_dir}{path to review project root.}
}
\value{
setup review project with templates
}
\description{
Create a review project using `pkgreview_create`. Wait till your workspace has
been switched to pkg review project dircetory and initialise review using
`pkgreview_init`
}
\examples{
\dontrun{
pkgreview_init(pkg_repo = "cboettig/rdflib")
}
}
