% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview.R
\name{pkgreview_create}
\alias{pkgreview_create}
\title{Create a review project.}
\usage{
pkgreview_create(
  pkg_repo,
  review_parent = ".",
  template = c("review", "editor"),
  issue_no = NULL
)
}
\arguments{
\item{pkg_repo}{character string of the repo owner and name in the form of
\code{"owner/repo"}.}

\item{review_parent}{directory in which to setup review project and source package
source code.}

\item{template}{character string, one of \code{review} or \code{editor}.}

\item{issue_no}{integer. Issue number of the pkg review in the rOpenSci \href{https://github.com/ropensci/software-review/issues}{\code{software-review} repository}.
If \code{NULL} (default), the issue number is extracted from the rOpenSci \strong{Under Review} badge on the pkg repository README.
Supplying an integer to \code{issue_no} overrides this behaviour and can be useful if a badge has not been added to the README yet.}
}
\value{
setup review project with templates
}
\description{
Create and initialise an rOpenSci package review project
}
\examples{
\dontrun{
# for a review project
pkgreview_create(pkg_repo = "ropensci/rdflib", review_parent = "~/Documents/reviews/")
# for editors checks
pkgreview_create(pkg_repo = "ropensci/rdflib", review_parent = "~/Documents/editorials/",
template = "editor")
}
}
