% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview.R
\name{pkgreview_getdata}
\alias{pkgreview_getdata}
\title{pkgreview_getdata}
\usage{
pkgreview_getdata(pkg_dir = NULL, pkg_repo, template = c("review",
  "editor"))
}
\arguments{
\item{pkg_dir}{path to the package source code directory}

\item{pkg_repo}{character string of the repo owner and name in the form of
\code{"owner/repo"}.}

\item{template}{character string, one of \code{review} or \code{editor}.}
}
\value{
a list of package metadata
}
\description{
get package metadata from package source code.
}
\examples{
\dontrun{
# run from within a pkgreviewr project with the package source code in a
sibling directory
pkgreview_getdata("../rdflib")
}
}
