% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-templates.R
\name{pkgreview_index_rmd}
\alias{pkgreview_index_rmd}
\alias{pkgreview_readme_md}
\alias{pkgreview_request}
\title{Create review templates.}
\usage{
pkgreview_index_rmd(pkg_data, template = c("review", "editor"), destdir)

pkgreview_readme_md(pkg_data, destdir)

pkgreview_request(pkg_data, destdir)
}
\arguments{
\item{pkg_data}{package metadata generated by pkgreview_getdata()}

\item{template}{character string, one of \code{review} or \code{editor}.}

\item{destdir}{where to save the template}
}
\description{
Creates skeleton review files:
\itemize{
\item \code{index.Rmd}: \code{html_notebook} to perform and record review in
\item \code{README.md}: prepopulated README for review repo.
}
\code{index.Rmd} will be automatically added to \code{.Rbuildignore}.
The resulting templates are populated with default
YAML frontmatter and R fenced code chunks (\code{Rmd}).
}
