% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview_init.R
\name{pkgreview_index_rmd}
\alias{pkgreview_index_rmd}
\alias{pkgreview_readme_md}
\alias{pkgreview_pkgreview_md}
\title{Create review templates.}
\usage{
pkgreview_index_rmd(pkg_data, open = interactive())

pkgreview_readme_md(pkg_data, open = interactive())

pkgreview_pkgreview_md(pkg_data, open = interactive())
}
\arguments{
\item{open}{allow interaction}

\item{pkg_dir}{package directory}
}
\description{
Creates skeleton review files:
\itemize{
\item `index.Rmd`: `html_notebook` to perform and record review in
\item `pkgreview.md`: review response submission template.
\item `README.md`: prepopulated README for review repo.
}
`index.Rmd` will be automatically added to `.Rbuildignore`. The resulting templates are populated with default
YAML frontmatter and R fenced code chunks (`Rmd`).
}
\examples{
\dontrun{
pkgreview_index_rmd("../rdflib/")
pkgreview_readme_md("../rdflib/")
pkgreview_pkgreview_md("../rdflib/")
}
}
