% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-templates.R
\name{pkgreview_index_rmd}
\alias{pkgreview_index_rmd}
\alias{pkgreview_readme_md}
\title{Create review templates.}
\usage{
pkgreview_index_rmd(pkg_data)

pkgreview_readme_md(pkg_data)
}
\arguments{
\item{pkg_data}{package metadata generated by pkgreview_getdata()}
}
\description{
Creates skeleton review files:
\itemize{
\item `index.Rmd`: `html_notebook` to perform and record review in
\item `README.md`: prepopulated README for review repo.
}
`index.Rmd` will be automatically added to `.Rbuildignore`. The resulting templates are populated with default
YAML frontmatter and R fenced code chunks (`Rmd`).
}
\examples{
\dontrun{
pkg_data <- pkgreview_getdata(pkg_dir)
pkgreview_index_rmd(pkg_data)
pkgreview_readme_md(pkg_data)
}
}
