% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview.R
\name{pkgreview_init}
\alias{pkgreview_init}
\title{Initialise pkgreview}
\usage{
pkgreview_init(pkg_repo, review_dir = NULL, pkg_dir = NULL, tmp = NULL)
}
\arguments{
\item{pkg_repo}{character string of the repo owner and name in the form of
`"owner/repo"`.}

\item{review_dir}{path to the review directory}

\item{pkg_dir}{path to package source directory, cloned from github}

\item{tmp}{used to specify a tmp directory during project creation.
Ignore for manual initialisation}
}
\value{
Initialisation creates pre-populated `index.Rmd`, `pkgreview.md` and `README.md` documents.
To initialise correctly, the function requires that the source code for the
package has been cloned. This might need to be done manually if it failed
during review creation. If setup is correct. Defaults are set to work in the
root of the `review_dir` with `pkg_dir` set to `"../`
}
\description{
Initialise pkgreview
}
\examples{
\dontrun{
pkgreview_init(pkg_repo = "cboettig/rdflib")
}
}
