% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreviewr-package.R
\docType{package}
\name{pkgreviewr-package}
\alias{pkgreviewr}
\alias{pkgreviewr-package}
\title{pkgreviewr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
