% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreviewr-package.R
\docType{package}
\name{pkgreviewr-package}
\alias{pkgreviewr}
\alias{pkgreviewr-package}
\title{pkgreviewr: rOpenSci package review project template}
\description{
Creates files and collects materials necessary to complete an rOpenSci package review.
Review files are prepopulated with review package specific metadata. Review package source code is
also cloned for local testing and inspection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/pkgreviewr}
  \item Report bugs at \url{https://github.com/ropenscilabs/pkgreviewr/issues}
}

}
\author{
\strong{Maintainer}: Anna Krystalli \email{annakrystalli@googlemail.com}

}
\keyword{internal}
