% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev_calls}
\alias{rev_calls}
\title{Create a dataframe of functions that a called and called by}
\usage{
rev_calls(path = ".")
}
\arguments{
\item{path}{Name of a source R package tar archive file,
or path to the folder of an R package.}
}
\value{
A dataframe of functions in and used by the package and the number of times they are called and called by
}
\description{
Create a dataframe of all functions in and used by the package
}
