% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_recursive.R
\name{rev_recursive}
\alias{rev_recursive}
\title{Review recursive function calls}
\usage{
rev_recursive(path = ".", igraph_obj = NULL)
}
\arguments{
\item{path}{Name of a source R package tar archive file,
or path to the folder of an R package.}

\item{igraph_obj}{igraph object for function calls dependencies returned by `create_package_igraph()`}
}
\value{
A two-column dataframe
}
\description{
`rev_recursive()` counts the number of functions that depend on each function in the package under review
}
\examples{
\dontrun{
rev_recursive(igraph_obj)
}
}
