% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_recursive.R
\name{rev_recursive}
\alias{rev_recursive}
\title{Review recursive function calls}
\usage{
rev_recursive(igraph_obj)
}
\arguments{
\item{igraph_obj}{object returned by `create_package_igraph()`}
}
\value{
A two-column dataframe
}
\description{
`rev_recursive()` counts the number of functions that depend on each function in the package under review, taking the object of class `igraph` returned by `create_package_igraph()` as input
}
\examples{
\dontrun{
rev_recursive(igraph_obj)
}
}
