% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-templates.R
\name{use_onboarding_tmpl}
\alias{use_onboarding_tmpl}
\title{Create onboarding template}
\usage{
use_onboarding_tmpl(template = c("review", "editor"))
}
\arguments{
\item{template}{character string, one of \code{review} or \code{editor}.}
}
\value{
writes a \code{pkg{template}.md} checklist template file in the project root.
}
\description{
Clone an up to date copy of the specified ropensci onboarding template.
}
